tui: Move file review candidate browser
The TUI file review flow now has a separate display module, while the main
file review module still owns file selection, navigation, and review action
routing.

Maintainers still have to read batch candidate preview and execution code
inside the main review loop module. That keeps batch-source transfer
details near unrelated live-source actions and makes the ownership boundary
harder to see.

This commit addresses that by moving batch candidate browsing into a
file_review.candidates module and routing the review loop through
browse_candidates.

Subsequent commits will add an architecture check for the new owner.
